Group custom SWR hooks with mutation methods
